Experiment of TRAINED vs EVALUATED vs LOADED
Tries finding the stitched network accuracy (still using that
abstraction) for a network that is getting its accuracy from the
evaluate loop inside the train_loop call vs. one which just uses a
stitch stored in memory (already trained) + evaluate (with the train
loader) vs one that is loaded from disk (that first `train_loop` table
loop stores). In theory they should ALL be the SAME. If they are not,
then something is super off.